[next-auth] jwt 사용자 지정 로그인 & register 회원가입 페이지 생성 & next-auth CredentialsProvider 생성

+++ pages/api/abc/[result].js
@@ -13,23 +13,41 @@ export default async function handler(req, res) {
 
     // dle1883@gmail.com
     // console.log(JSON.parse(req.query.result).author);
-    
+
     // dle1883@gmail.com
     // console.log(session.user.email)
 
-    
+
     // console.log(JSON.parse(req.query.result).author)
     if (req.method === 'DELETE') {
-        // 터미널 error창 session 부분 에러
-        if (JSON.parse(req.query.result).author === session.user.email) {
-            console.log('일치')
-            try {
-                const result = await db.collection('post').deleteOne({ _id: new ObjectId(JSON.parse(req.query.result)._id)})
-                res.status(200).json('삭제완료')
-            } catch (error) {
-                console.log(error);
+        // *******터미널 error창 session 부분 에러 확인******
+
+        if (session) {
+            if (JSON.parse(req.query.result).author === session.user.email) {
+                console.log('일치')
+                try {
+                    const result = await db.collection('post').deleteOne({ _id: new ObjectId(JSON.parse(req.query.result)._id) })
+                    res.status(200).json('삭제완료')
+                } catch (error) {
+                    console.log(error);
+                }
+            } else {
+                // console.log('작성자가 아닙니다.')
+                return res.status(500).json('작성자가 아닙니다.')
             }
-        } 
+        } else {
+            res.status(500).json('로그인 먼저 부탁')
+        }
+
+        // if (JSON.parse(req.query.result).author === session.user.email) {
+        //     console.log('일치')
+        //     try {
+        //         const result = await db.collection('post').deleteOne({ _id: new ObjectId(JSON.parse(req.query.result)._id)})
+        //         res.status(200).json('삭제완료')
+        //     } catch (error) {
+        //         console.log(error);
+        //     }
+        // }
+
     }
 }
-
